dm: block: fix init failure of a block device with discard
When boot vm configured a virtio-blk device with discard, an error
occurs:

```
acrn_dm: not support DISCARD
```

`BLKDISCARD` ioctl is invoked with parameter {0,0} while initialization.
This is used to check the ioctl capability. But kernel will return error
if {0,0}.

There is no specific interface to check capability in kernel. So save
the discard state in blockif_ctxt->failed_discard to check before
process. And give one more chance while reset.

Tracked-On: OAM-118078
Signed-off-by: Haiwei Li <haiwei.li@intel.com>