misc: Add Support in configtool to generate virtio-camera script
This patch is to generate a json file and
update launch script to support virtio-camera.
I formatted scenario data and produce a specific json file.
I let the allocator to dispatch one bdf to camera.

I run the scripts locally, changed the scenario.xml several times
and observated the products are all correct.

Signed-off-by: dongpingx <dongpingx.wu@intel.com>
Tracked-On: ACRN-10526