fix: fix the file name of disk logger
we init the logger through cmdline, and
also create a file to store the acrn-dm log msg.

But the logger filename is usually wrong, now it's
/var/log/acrn-dm/\(null\)_log_0
The reason is that we use vmname to set the log
filename. When the vmname still is NULL,
will create a wrong log file name.

So, after we parse the cmdline, the vmname should be set
immediately. And following pr_xxx() can create
the appropriate log file.

Tracked-On: ACRN-10722
Signed-off-by: Zhangwei6 <wei6.zhang@intel.com>