dm: virtio-camera set fmt for ipu
Due to the current implementation of ipu6 hal
and driver, the fmt information will be lost
when stream off a camera stream. This behavior
does not conform to normal camera devices,
such as USB camera.

So we add additional operations for the ipu.
Which is to set_fmt after stream_off to
simulate a normal camera device.

Tracked-On: ACRN-10843
Signed-off-by: Zhangwei6 <wei6.zhang@intel.com>